chore: Move nonlinear models to separate file
